requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
pdfplumber==0.11.4
gtts==2.5.4
pyttsx3==2.99
httpx==0.28.1
pytest==8.3.4
openai==2.21.0
google-genai==1.63.0
tenacity==9.1.4
websockets==15.0.1